Refactored Lemonings to use NavMeshPath instead of NavMeshAgent
Trying to get an Animator to play nicely with NavMeshAgent and a
RigidBody turned out to be an exercise in futility. The agent kept
getting desynced with the model and hitbox. Some implementations
would get the character stuck in the floor, or even fall through.

The problem is that agents will always move to follow its path,
regardless of everything else, causing a desync. Trying to force
the body to follow the agent causes wonky physics, hence falling
through the floor.